Adds JEST lint, new tests, readme, travis and test coverage
Tests weren't linted which is now fixed with eslint-plugin-jest.
Func allowed non called function which makes very little sense, so it was removed.
It is now possible to call Func with a message that will automatically
tranform the error if there is any.
Travis-ci is now configured and should run the coverage after the test and send results to coveralls.